+++ Scripts/Runtimes/Enums/AxisEventType.cs
@@ -1,0 +1,10 @@
+﻿namespace DSC.Core
+{
+    public enum AxisEventType
+    {
+        Press           =   1 << 0,
+        DoublePress     =   1 << 1,
+        Tap             =   1 << 2,
+        DoubleTap       =   1 << 3
+    }
+}