+++ Scripts/Runtimes/Enums/AxisType.cs
@@ -1,0 +1,8 @@
+﻿namespace DSC.Core
+{
+    public enum AxisType
+    {
+        Horizontal  =   1 << 0,
+        Vertical    =   1 << 1
+    }
+}